修改gitignore

+++ app/src/main/java/com/victor/myclient/view/Case/CaseListActivity.java
@@ -7,7 +7,6 @@ import android.os.Handler;
 import android.os.Message;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.View;
 import android.widget.RelativeLayout;
 
@@ -22,6 +21,7 @@ import com.victor.myclient.view.Case.CaseLayout.CustomLayoutManager;
 
 import org.litepal.crud.DataSupport;
 
+import java.lang.reflect.Type;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -32,15 +32,12 @@ import demo.animen.com.xiaoyutask.R;
  */
 
 public class CaseListActivity extends AppCompatActivity {
-
     public RecyclerView recyclerView;
     private RelativeLayout back;
     private String patientId;
-
-    private List<CaseInfor> caseInfors;
+    private List<CaseInfor> caseInfors = new ArrayList<>();
     private CaseAdapter adapter;
     private ProgressDialog progressDialog;
-
     private boolean net_work_available, has_data;
     private Handler handler = new Handler() {
         @Override
@@ -50,7 +47,7 @@ public class CaseListActivity extends AppCompatActivity {
             }
         }
     };
-
+    @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.case_list);
@@ -62,10 +59,9 @@ public class CaseListActivity extends AppCompatActivity {
         } else {
             initView();
             initEvent();
-            new CaseListTask().execute(patientId);
+            new CaseListTask().execute();
         }
     }
-
     private void initView() {
         back = (RelativeLayout) findViewById(R.id.case_list_back);
         recyclerView = (RecyclerView) findViewById(R.id.recycler_view);
@@ -77,6 +73,8 @@ public class CaseListActivity extends AppCompatActivity {
         progressDialog.show();
     }
 
+
+
     private void initEvent() {
         back.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -86,12 +84,11 @@ public class CaseListActivity extends AppCompatActivity {
         });
     }
 
-
-    class CaseListTask extends AsyncTask<String, Void, String> {
+    private class CaseListTask extends AsyncTask<String, Void, String> {
         private Gson gson = new Gson();
 
-        @Override
 
+        @Override
         protected void onPostExecute(String s) {
             super.onPostExecute(s);
             if (s != null && has_data) {
@@ -108,15 +105,18 @@ public class CaseListActivity extends AppCompatActivity {
         @Override
         protected String doInBackground(String... params) {
             if (net_work_available) {
-                caseInfors = gson.fromJson(Utils.sendRequest(GlobalData.GET_PATIENT_CASE + patientId), new TypeToken<List<CaseInfor>>() {
-                }.getType());
-                DataSupport.deleteAll(CaseInfor.class);
-                for (CaseInfor caseInfor : caseInfors) {
-                    if (!caseInfor.isSaved()) {
-                        caseInfor.save();
+                String info = Utils.sendRequest(GlobalData.GET_PATIENT_CASE + patientId);
+                if (!info.contains("not_exist")) {
+                    caseInfors = gson.fromJson(info, new TypeToken<List<CaseInfor>>() {
+                    }.getType());
+                    DataSupport.deleteAll(CaseInfor.class);
+                    for (CaseInfor caseInfor : caseInfors) {
+                        if (!caseInfor.isSaved()) {
+                            caseInfor.save();
+                        }
                     }
+                    has_data = true;
                 }
-                has_data = true;
 
             } else {
                 if (DataSupport.isExist(CaseInfor.class)) {
